Skeleton for historical prices module and added candle chart.

+++ Components.LivePrices/Models/Instrument.cs
@@ -4,12 +4,12 @@ namespace DeepInsights.Components.LivePrices.Models
 {
     public abstract class Instrument
     {
-        public double Spread { get; set; }
+        public decimal Spread { get; set; }
         public string QuoteName { get; set; }
-        public double Bid { get; set; }
-        public double Ask { get; set; }
-        public double LowestBid { get; set; }
-        public double HighestAsk { get; set; }
+        public decimal Bid { get; set; }
+        public decimal Ask { get; set; }
+        public decimal LowestBid { get; set; }
+        public decimal HighestAsk { get; set; }
 
     }
 }
